CoulombCollisions: More Adjustability + more info
This commit increases the adjustability of the Coulomb collision model.
User are now able to specify the species which are include in the calculation of the Debye length.
Also the consideration of bmin in the calculation of b0 can be turned of at runtime.
Lastly the user is now able to decide whether to include the drift velocity or not in the calculation
of the dirft velocity, which in turn is required for the calculation of the Debye length.

Now there is also more info about the collisions printed to the log.
This includes the Average Debye length, the average impact parameter b0, and the average Coulomb logartihm.