PairMerging: Add cell maximum
With this addition the user is able to specify an maximum of parcels
that can be merged in one cell.
Also the warning about untested code is removed. This model has been
successfully tested beam instability simulations. However user should
still take care with the settings in this model as to much merging can
cause huge errors.