CoulombCollision: Check if the Debye length has to be calculated
If the user provides a Coulomb logarithm for all possible collisions we can skip the calculation of the
Debye length. Now we correclty check only for pairs of charged species and not all species.
Note that the statistics logging output will print zero for this quantity.
This commit also makes some init messages more pretty.